Redoculous.js now simply does html modification instead of pushstate, original moved to redoculous-pushstate.js

+++ app/src/main/java/org/ocpsoft/redoculous/util/Files.java
@@ -109,7 +109,7 @@ public final class Files
             {
                _deleteRecursive(sf, false);
             }
-            else
+            else if (sf.exists())
             {
                if (!sf.delete())
                {
